+++ packages/at_client/test/pairwise_secret_sharing_test.dart
@@ -122,7 +122,7 @@ void main() {
     final atLookUp = MockAtLookUp();
     when(() => atClient.getRemoteSecondary()).thenReturn(remoteSecondary);
     when(() => remoteSecondary.atLookUp).thenReturn(atLookUp);
-    when(() => atLookUp.enrollmentId).thenReturn(enrollmentId);
+    when(() => atClient.enrollmentId).thenReturn(enrollmentId);
 
     when(() => atClient.put(any(), any(),
         putRequestOptions: any(named: 'putRequestOptions'))).thenAnswer((inv) {